models change & reset_db tested ok
--amend

+++ app/utils/reset_db.py
@@ -1,23 +1,48 @@
 import os
 import csv
+from typing import Optional
 
+from loguru import logger
 from sqlalchemy.orm import Session
 from sqlalchemy.engine import Engine
+from sqlalchemy.schema import CreateTable, DropTable
 
+from app.constant.confirm import confirms_execution
 from app.db.engine import engine_from_env
 from app.db.models import Market, MetadataBase
 
 
-MARKET_CSV_FILE = os.path.join(os.path.dirname(__file__), "../db/", 'market.csv')
+MARKET_CSV_FILE = os.path.join(os.path.dirname(__file__), "../constant/", 'market.csv')
 
 
-def reset_table_content(engine: Engine, reset=False):
-    with Session(engine) as session:
-        if reset:
+def reset_table_content(
+    engine: Engine, 
+    dryrun: Optional[bool] = False,
+    reset: Optional[bool] = False,
+    yes: Optional[bool] = False,
+):
+
+    if reset:
+        if dryrun:
+            for table in MetadataBase.metadata.tables.values():
+                logger.info(DropTable(table).compile(dialect=engine.dialect))
+        else:
+            confirms_execution(
+                f"Resetting tables in {engine.url.database} at {engine.url.host}",
+                defaultY=False,
+                yes=yes
+            )
             MetadataBase.metadata.drop_all(engine)
 
-        # CREATE TABLE :name IF NOT EXISTS
-        MetadataBase.metadata.create_all(engine)
+    with Session(engine) as session:
+        if dryrun:
+            for table in MetadataBase.metadata.tables.values():
+                logger.info(CreateTable(table).compile(dialect=engine.dialect))
+        
+        else:    
+            # equiv. to
+            # CREATE TABLE :name IF NOT EXISTS
+            MetadataBase.metadata.create_all(engine)
 
         # preload market info
         with open(MARKET_CSV_FILE, newline='') as csvfile:
@@ -25,10 +50,14 @@ def reset_table_content(engine: Engine, reset=False):
             for row in reader:
                 market = Market(**row)
                 session.add(market)
-            session.commit()
+            
+            session.commit() if not dryrun else session.rollback()
 
 
 if __name__ == '__main__':
-    reset_input = input("Do you want to reset the database? (y/[n]): ")
-    reset = True if reset_input.lower() == 'y' else False
-    reset_table_content(engine_from_env(), reset=reset)
+    reset_table_content(
+        engine_from_env(echo=True), 
+        dryrun=False,
+        reset=True, 
+        yes=False,
+    )
